Add a Root Layout Component to the web-core
