+++ academy-ui/src/app/page.tsx
@@ -10,7 +10,7 @@ import TidbitsHubHome from '@/components/home/TidbitsHub/TidbitsHubHome';
 import ListProjects from '@/components/projects/List/ListProjects';
 import { ProjectFragment, SpaceTypes } from '@/graphql/generated/generated-types';
 import getApiResponse from '@/utils/api/getApiResponse';
-import { getSpaceServerSide } from '@/utils/api/getSpaceServerSide';
+import { getSpaceServerSide } from '@dodao/web-core/api/auth/getSpaceServerSide';
 import { authOptions } from '@/app/api/auth/[...nextauth]/authOptions';
 import { getServerSession } from 'next-auth';
 import { headers } from 'next/headers';
