+++ academy-ui/src/app/projects/view/[projectId]/layout.tsx
@@ -2,7 +2,7 @@ import PageWrapper from '@dodao/web-core/components/core/page/PageWrapper';
 import { ViewProjectHeader } from '@/components/projects/View/ViewProjectHeader';
 import { ProjectFragment } from '@/graphql/generated/generated-types';
 import getApiResponse from '@/utils/api/getApiResponse';
-import { getSpaceServerSide } from '@/utils/api/getSpaceServerSide';
+import { getSpaceServerSide } from '@dodao/web-core/api/auth/getSpaceServerSide';
 import React from 'react';
 
 async function ProjectViewHome(props: { params: { projectId: string; viewType: string }; children: React.ReactNode }) {
