+++ academy-ui/src/app/shorts/page.tsx
@@ -1,7 +1,7 @@
 import Shorts from '@/components/shortVideos/View/Shorts';
 import { ProjectShortVideo, ShortVideo, ShortVideoFragment } from '@/graphql/generated/generated-types';
 import getApiResponse from '@/utils/api/getApiResponse';
-import { getSpaceServerSide } from '@/utils/api/getSpaceServerSide';
+import { getSpaceServerSide } from '@dodao/web-core/api/auth/getSpaceServerSide';
 import { Metadata } from 'next';
 
 export const metadata: Metadata = {
