+++ shared/web-core/src/api/auth/createHash.ts
